finished

+++ mercadinho/src/br/edu/ecit/primeiroprojeto/mercadinho/modelo/produto/Produto.java
@@ -21,7 +21,7 @@ public abstract class Produto {
         this.valor = valor;
     }
 
-    public long getCodigo() {
+    public int getCodigo() {
         return codigo;
     }
